Change Float32 to Float64

+++ src/estimate/smc/smc.jl
@@ -265,7 +265,7 @@ function smc(m::AbstractModel, data::T;
 
     if !m.testing
         simfile = h5open(rawpath(m, "estimate", "smcsave.h5"), "w")
-        particle_store = d_create(simfile, "smcparams", datatype(Float32),
+        particle_store = d_create(simfile, "smcparams", datatype(Float64),
                                   dataspace(n_parts, n_params))
         for i in 1:length(cloud)
             particle_store[i,:] = cloud.particles[i].value
